requirements.txt
# DO NOT EDIT!
# Generate by running 'bazel run --config=pre_build //bazel/requirements:sync_requirements'

--extra-index-url https://pypi.org/simple
absl-py==1.3.0
accelerate==0.22.0
anyio==3.5.0
boto3==1.24.28
cachetools==4.2.2
cloudpickle==2.0.0
coverage==6.3.2
cryptography==39.0.1
flask-cors==3.0.10
flask==2.1.3
fsspec[http]==2023.3.0
httpx==0.23.0
importlib_resources==5.1.4
inflection==0.5.1
joblib==1.1.1
jsonschema==3.2.0
lightgbm==3.3.5
mlflow==2.3.1
moto==4.0.11
mypy==1.5.1
networkx==2.8.4
numpy==1.24.3
packaging==23.0
pandas==1.5.3
peft==0.5.0
protobuf==3.20.3
pyarrow==10.0.1
pytest==7.4.0
pytimeparse==1.1.8
pyyaml==6.0
retrying==1.3.3
ruamel.yaml==0.17.21
s3fs==2023.3.0
scikit-learn==1.3.0
scipy==1.9.3
sentencepiece==0.1.99
shap==0.42.1
snowflake-connector-python[pandas]==3.2.0
snowflake-snowpark-python==1.8.0
sphinx==5.0.2
sqlparse==0.4.4
starlette==0.27.0
tensorflow==2.13.1
tokenizers==0.13.2
torch==2.0.1
torchdata==0.6.1
transformers==4.32.1
types-PyYAML==6.0.12
types-cachetools==4.2.2
types-protobuf==4.23.0.1
types-requests==2.30.0.0
typing-extensions==4.5.0
xgboost==1.7.3
